api/cinemAPI/requirements.txt
asgiref==3.4.1
Django==3.1.12
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangotoolbox==1.8.0
djongo==1.3.6
jsonfield==3.1.0
psycopg2-binary==2.9.1
pymongo==3.12.0
pytz==2021.1
sqlparse==0.2.4
